Update training configuration in config.yaml for improved performance; change context size to 64 and online steps to 1024. Modify train.py to use float16 for mixed precision training.

+++ train.py
@@ -165,7 +165,7 @@ def train_on_task(
 
         # ----- forward pass under mixed precision -------------------
         with (
-            autocast(device_type="cuda", dtype=torch.float32)
+            autocast(device_type="cuda", dtype=torch.float16)
             if TORCH_DEVICE == "cuda"
             else contextlib.nullcontext()
         ):
